+++ lib/src/models/request/a2a_request_header.g.dart
@@ -13,7 +13,6 @@ _PPGeneralRequestHeader _$PPGeneralRequestHeaderFromJson(
     _$PPTransactionTypeEnumMap,
     json['transaction_type'],
   ),
-  clientToken: json['client_token'] as String,
   orderCode: json['order_code'] as String?,
   transactionId: json['transaction_id'] as String?,
 );
@@ -22,7 +21,6 @@ Map<String, dynamic> _$PPGeneralRequestHeaderToJson(
   _PPGeneralRequestHeader instance,
 ) => <String, dynamic>{
   'transaction_type': _$PPTransactionTypeEnumMap[instance.transactionType]!,
-  'client_token': instance.clientToken,
   'order_code': instance.orderCode,
   'transaction_id': instance.transactionId,
 };