+++ lib/src/models/request/a2a_start_payment_request.dart
@@ -17,8 +17,8 @@ sealed class PPStartPaymentRequestModel with _$PPStartPaymentRequestModel {
       _$PPStartPaymentRequestModelFromJson(json);
 
   factory PPStartPaymentRequestModel.toRequest({
-    
-    required String clientToken,
+    // clientToken: temporarily removed. May be re-enabled in the future.
+    // required String clientToken,
     required String orderCode,
     required PPPaymentType paymentType,
     required PPPaymentMethod paymentMethod,
@@ -41,7 +41,7 @@ sealed class PPStartPaymentRequestModel with _$PPStartPaymentRequestModel {
     ),
     header: PPGeneralRequestHeader(
       transactionType: PPTransactionType.POST_PAYMENT_START,
-      clientToken: clientToken,
+      // clientToken: clientToken,
       orderCode: orderCode,
     ),
   );